% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms.R
\name{dimsumms}
\alias{dimsumms}
\title{dimsumms}
\usage{
dimsumms(startStage = 1, stopStage = 0,
  base_dir = "/users/blehner/afaure/DMS/Results/dimsumms_proj")
}
\arguments{
\item{startStage}{Start at a specified analysis stage (default:1)}

\item{stopStage}{Stop at a specified analysis stage (default:0 i.e. no stop condition)}

\item{base_dir}{Base directory for all output file (default:NB private CRG server path; change accordingly)}
}
\value{
Nothing
}
\description{
Main analysis script.
}
