% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms__gg_color_hue.R
\name{dimsumms__gg_color_hue}
\alias{dimsumms__gg_color_hue}
\title{dimsumms__gg_color_hue}
\usage{
dimsumms__gg_color_hue(n)
}
\arguments{
\item{n}{an integer number of colours required (required)}
}
\value{
a vector of colours
}
\description{
ggplot2-like colour scale in HCL space.
}
