% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms__replicate_bottleneck.R
\name{dimsumms__replicate_bottleneck}
\alias{dimsumms__replicate_bottleneck}
\title{dimsumms__replicate_bottleneck}
\usage{
dimsumms__replicate_bottleneck(input_file, outpath, bottleneck.alpha,
  Sequencing.Error = 0.02)
}
\arguments{
\item{input_file}{path to input file (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{bottleneck.alpha}{bottleneck factor}

\item{Sequencing.Error}{sequencing error (default:0.02)}
}
\value{
Nothing
}
\description{
Simulate a replicate bottleneck.
}
