% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_aa_properties_mutant_effects.R
\name{dimsumms_aa_properties_mutant_effects}
\alias{dimsumms_aa_properties_mutant_effects}
\title{dimsumms_aa_properties_mutant_effects}
\usage{
dimsumms_aa_properties_mutant_effects(toxicity_dt, outpath, aaprop_file,
  aaprop_file_selected, colour_scheme, report = F)
}
\arguments{
\item{toxicity_dt}{data.table with single and double mutant codes (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{aaprop_file}{path to amino acid properties file (required)}

\item{aaprop_file_selected}{path to file with selected subset of identifiers}

\item{colour_scheme}{colour scheme file (required)}

\item{report}{whether or not to save output and plots (default:F)}
}
\value{
A data.table with single and double mutant effects from AA PCA
}
\description{
Calculate single and double mutant effects from AA PCA (using single mutants).
}
