% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_bottleneck_fitness_scatter.R
\name{dimsumms_bottleneck_fitness_scatter}
\alias{dimsumms_bottleneck_fitness_scatter}
\title{dimsumms_bottleneck_fitness_scatter}
\usage{
dimsumms_bottleneck_fitness_scatter(fitness_list, outpath,
  execute = TRUE)
}
\arguments{
\item{fitness_list}{list of data.tables (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot fitness before/after bottleneck and different minimum input count thresholds
}
