% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dimsumms_bottleneck_fitness_vs_hydrophobicity_scatter.R
\name{dimsumms_bottleneck_fitness_vs_hydrophobicity_scatter}
\alias{dimsumms_bottleneck_fitness_vs_hydrophobicity_scatter}
\title{dimsumms_bottleneck_fitness_vs_hydrophobicity_scatter}
\usage{
dimsumms_bottleneck_fitness_vs_hydrophobicity_scatter(
  fitness_list,
  outpath,
  aaprop_file,
  aaprop_file_selected,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{fitness_list}{list of data.tables (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{aaprop_file}{path to amino acid properties file (required)}

\item{aaprop_file_selected}{path to file with selected subset of identifiers}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot fitness before/after bottleneck and different minimum input count thresholds
}
