% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_bottleneck_simulations.R
\name{dimsumms_bottleneck_simulations}
\alias{dimsumms_bottleneck_simulations}
\title{dimsumms_bottleneck_simulations}
\usage{
dimsumms_bottleneck_simulations(dataset_dir, execute = TRUE)
}
\arguments{
\item{dataset_dir}{directory with pre-processed DMS data}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Simulate bottlenecks
}
