% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_error_model_performance.R
\name{dimsumms_error_model_performance}
\alias{dimsumms_error_model_performance}
\title{dimsumms_error_model_performance}
\usage{
dimsumms_error_model_performance(performance_dir, outpath,
  execute = TRUE)
}
\arguments{
\item{performance_dir}{Base directory with error model performance results}

\item{outpath}{output path for plots and saved objects (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot error model performance results
}
