% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_errormodel_data_manipulations.R
\name{dimsumms_errormodel_data_manipulations}
\alias{dimsumms_errormodel_data_manipulations}
\title{dimsumms_errormodel_data_manipulations}
\usage{
dimsumms_errormodel_data_manipulations(
  error_model_dir,
  outpath,
  execute = TRUE
)
}
\arguments{
\item{error_model_dir}{Base directory with subfolders containing results from different DiMSum runs}

\item{outpath}{output path for plots and saved objects (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plot error model parameters before/after replicate error and over-sequencing factor manipulations in TDP43 290 library
}
