% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_errormodel_fit.R
\name{dimsumms_errormodel_fit}
\alias{dimsumms_errormodel_fit}
\title{dimsumms_errormodel_fit}
\usage{
dimsumms_errormodel_fit(DT, reps, Ncores = 1, Nbootstraps = 100,
  maxN = 10000, max_tries_per_fit = 20, lower_bound_add = 10^-4,
  Fcorr = c())
}
\arguments{
\item{DT}{DMS dataset data.table}

\item{reps}{replicates for which the error model parameters should be estimated}

\item{Ncores}{how many cores to use for parallelization}

\item{Nbootstraps}{# bootstraps for error model calculation}

\item{maxN}{how many variants are used per bootstrap}

\item{max_tries_per_fit}{how many restarts of the fitting routine in case nls doesn't yield results}

\item{lower_bound_add}{lower limited for additive error terms}

\item{Fcorr}{fitness normalization parameters}
}
\value{
Nothing
}
\description{
estimate multiplicative input/output and additive error terms from DMS dataset
}
