% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dimsumms_errormodel_leaveoneout_preprocess_datasets.R
\name{dimsumms_errormodel_leaveoneout_preprocess_datasets}
\alias{dimsumms_errormodel_leaveoneout_preprocess_datasets}
\title{dimsumms_errormodel_leaveoneout_preprocess_datasets}
\usage{
dimsumms_errormodel_leaveoneout_preprocess_datasets(
  dataset_dir,
  execute = TRUE
)
}
\arguments{
\item{dataset_dir}{base directory with downloaded DiMSum manuscript data (folder 'misc')}

\item{execute}{whether or not to execute the analysis (default: TRUE)}

\item{outpath}{output path for processed datasets}
}
\value{
Nothing
}
\description{
Preprocess datasets for leave one out cross validation benchmarking
}
