% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_errormodel_random_effect.R
\name{dimsumms_errormodel_random_effect}
\alias{dimsumms_errormodel_random_effect}
\title{dimsumms_errormodel_random_effect}
\usage{
dimsumms_errormodel_random_effect(score_rx, sigma_rx)
}
\arguments{
\item{score_rx}{replicate fitness scores (#score_rx is GxR matrix with G number of genotypes, R number of replicates)}

\item{sigma_rx}{replicate count-based errors}
}
\value{
Nothing
}
\description{
estimate variant specific random error terms
}
