% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_hierarchical_abundance.R
\name{dimsumms_hierarchical_abundance}
\alias{dimsumms_hierarchical_abundance}
\title{dimsumms_hierarchical_abundance}
\usage{
dimsumms_hierarchical_abundance(outpath, execute = TRUE)
}
\arguments{
\item{outpath}{output path for plots and saved objects (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Plots showing hierarchical abundance of DMS experiments
}
