% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_prepare_all_dms_dt.R
\name{dimsumms_prepare_all_dms_dt}
\alias{dimsumms_prepare_all_dms_dt}
\title{dimsumms_prepare_all_dms_dt}
\usage{
dimsumms_prepare_all_dms_dt(fitness_dir)
}
\arguments{
\item{fitness_dir}{Base directory with subfolders containing results from different DiMSum runs}
}
\value{
Nothing
}
\description{
Load and reformat a list of fitness data.tables.
}
