% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_prepare_dms_dt.R
\name{dimsumms_prepare_dms_dt}
\alias{dimsumms_prepare_dms_dt}
\title{dimsumms_prepare_dms_dt}
\usage{
dimsumms_prepare_dms_dt(fitness_list)
}
\arguments{
\item{file_path}{named list of folder paths with fitness estimates (required)}
}
\value{
Data.table with fitness estimates
}
\description{
Load and reformat fitness data.table.
}
