% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_real_bottleneck_scatterplot_matrices.R
\name{dimsumms_real_bottleneck_scatterplot_matrices}
\alias{dimsumms_real_bottleneck_scatterplot_matrices}
\title{dimsumms_real_bottleneck_scatterplot_matrices}
\usage{
dimsumms_real_bottleneck_scatterplot_matrices(
  dataset_dir,
  outpath,
  execute = TRUE
)
}
\arguments{
\item{dataset_dir}{directory with preprocessed DMS datasets (preprocessed by dimsumms_errormodel_leaveoneout_preprocess_datasets.R script)}

\item{outpath}{output path for plots and saved objects (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Scatterplot matrices of Input and Output sample variant counts for doubles from real DMS datasets
}
