% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms__library_bottleneck.R,
%   R/dimsumms_generate_bottlenecks.R
\name{dimsumms__library_bottleneck}
\alias{dimsumms__library_bottleneck}
\title{dimsumms__library_bottleneck}
\usage{
dimsumms__library_bottleneck(
  input_file,
  outpath,
  Bottleneck.Factor = 0.01,
  Sequencing.Error = 0.02
)

dimsumms__library_bottleneck(
  input_file,
  outpath,
  Bottleneck.Factor = 0.01,
  Sequencing.Error = 0.02
)
}
\arguments{
\item{input_file}{path to input file (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{Bottleneck.Factor}{bottleneck factor (default:0.01)}

\item{Sequencing.Error}{sequencing error (default:0.02)}
}
\value{
Nothing

Nothing
}
\description{
Simulate a library bottleneck.

Simulate a library bottleneck.
}
