% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimsumms_errormodel_leaveoneout.R
\name{dimsumms_errormodel_leaveoneout}
\alias{dimsumms_errormodel_leaveoneout}
\title{dimsumms_errormodel_leaveoneout}
\usage{
dimsumms_errormodel_leaveoneout(dataset_dir, Ncores = 1, outpath,
  execute = TRUE)
}
\arguments{
\item{dataset_dir}{directory with preprocessed DMS datasets (preprocessed by dimsumms_errormodel_leaveoneout_preprocess_datasets.R script)}

\item{Ncores}{number of cores to use to parallelize up analyes}

\item{outpath}{output path for plots and saved objects (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Perform leave one out cross validation to benchmark error models
}
